Cleanup Systemd unit file
Remove references to basic.target. Systemd automatically adds
dependencies of the types Requires= and After= for this target unit to
all services.

Remove KillMode= option. Consul itself only runs one process, and if it
spawns processes, e.g. for health checks, we do not want them to be left
behind when the service is stopped. The default is to kill all processes
in the control group.

Remove RestartSec= option. There is no apparent reason to override the
default of 100ms.